# pkg/api/zone/zone.gen.go
// Package zone provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package zone

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	"github.com/quanNaNbk03/monitor-demo/internal/server/dto"
)

// CreateZoneRequest defines model for CreateZoneRequest.
type CreateZoneRequest struct {
	IsTemplate     bool   `json:"isTemplate"`
	Name           string `json:"name" validate:"required,fqdn"`
	OrganizationID uint64 `json:"organizationID"`
	SoaType        string `json:"soaType"`
	Type           string `json:"type"`
}

// Error defines model for Error.
type Error struct {
	Error struct {
		Code    string `json:"code,omitempty"`
		Detail  string `json:"detail,omitempty"`
		Message string `json:"message,omitempty"`
		Name    string `json:"name,omitempty"`
	} `json:"error,omitempty"`
}

// ListZoneResponse defines model for ListZoneResponse.
type ListZoneResponse = dto.ListZoneResponse

// Paging defines model for Paging.
type Paging = dto.Paging

// PagingRequest defines model for PagingRequest.
type PagingRequest struct {
	Limit int `json:"limit,omitempty"`
	Page  int `json:"page,omitempty"`
}

// SearchRequest defines model for SearchRequest.
type SearchRequest struct {
	Field string      `json:"field,omitempty"`
	Op    string      `json:"op,omitempty"`
	Value interface{} `json:"value,omitempty"`
}

// UpdateZoneRequest update zone for admin
type UpdateZoneRequest struct {
	Dnssec  *bool   `json:"dnssec,omitempty"`
	SoaType *string `json:"soaType,omitempty"`
	Type    *string `json:"type,omitempty"`
}

// Zone defines model for Zone.
type Zone = dto.Zone

// ErrorResponse400 defines model for ErrorResponse400.
type ErrorResponse400 = Error

// ErrorResponse401 defines model for ErrorResponse401.
type ErrorResponse401 = Error

// ErrorResponse403 defines model for ErrorResponse403.
type ErrorResponse403 = Error

// ErrorResponse404 defines model for ErrorResponse404.
type ErrorResponse404 = Error

// ErrorResponse500 defines model for ErrorResponse500.
type ErrorResponse500 = Error

// ListZonesJSONBody defines parameters for ListZones.
type ListZonesJSONBody struct {
	Paging   PagingRequest   `json:"paging,omitempty"`
	Searches []SearchRequest `json:"searches,omitempty"`
}

// CreateJSONRequestBody defines body for Create for application/json ContentType.
type CreateJSONRequestBody = CreateZoneRequest

// ListZonesJSONRequestBody defines body for ListZones for application/json ContentType.
type ListZonesJSONRequestBody ListZonesJSONBody

// UpdateJSONRequestBody defines body for Update for application/json ContentType.
type UpdateJSONRequestBody = UpdateZoneRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create Create new zone
	// (POST /api/v1/zones)
	Create(c *gin.Context)
	// ListZones List all Zones
	// (POST /api/v1/zones/search)
	ListZones(c *gin.Context)
	// Delete Delete zone
	// (DELETE /api/v1/zones/{id})
	Delete(c *gin.Context, id uint64)
	// Get Get zone detail
	// (GET /api/v1/zones/{id})
	Get(c *gin.Context, id uint64)
	// Update Update zone
	// (PUT /api/v1/zones/{id})
	Update(c *gin.Context, id uint64)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// Create operation middleware
func (siw *ServerInterfaceWrapper) Create(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Create(c)
}

// ListZones operation middleware
func (siw *ServerInterfaceWrapper) ListZones(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ListZones(c)
}

// Delete operation middleware
func (siw *ServerInterfaceWrapper) Delete(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id uint64

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "uint64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Delete(c, id)
}

// Get operation middleware
func (siw *ServerInterfaceWrapper) Get(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id uint64

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "uint64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Get(c, id)
}

// Update operation middleware
func (siw *ServerInterfaceWrapper) Update(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id uint64

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "uint64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Update(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/api/v1/zones", wrapper.Create)
	router.POST(options.BaseURL+"/api/v1/zones/search", wrapper.ListZones)
	router.DELETE(options.BaseURL+"/api/v1/zones/:id", wrapper.Delete)
	router.GET(options.BaseURL+"/api/v1/zones/:id", wrapper.Get)
	router.PUT(options.BaseURL+"/api/v1/zones/:id", wrapper.Update)
}