# pkg/api/zoneclient/zone.gen.go
// Package zone provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package zone

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	"github.com/quanNaNbk03/monitor-demo/internal/server/dto"
)

// Error defines model for Error.
type Error struct {
	Error struct {
		Code    string `json:"code,omitempty"`
		Detail  string `json:"detail,omitempty"`
		Message string `json:"message,omitempty"`
		Name    string `json:"name,omitempty"`
	} `json:"error,omitempty"`
}

// ListZoneResponse defines model for ListZoneResponse.
type ListZoneResponse = dto.ListZoneResponse

// Paging defines model for Paging.
type Paging = dto.Paging

// PagingRequest defines model for PagingRequest.
type PagingRequest struct {
	Limit int `json:"limit,omitempty"`
	Page  int `json:"page,omitempty"`
}

// SearchRequest defines model for SearchRequest.
type SearchRequest struct {
	Field string      `json:"field,omitempty"`
	Op    string      `json:"op,omitempty"`
	Value interface{} `json:"value,omitempty"`
}

// UpdateZoneClientRequest update zone for client
type UpdateZoneClientRequest struct {
	Dnssec *bool `json:"dnssec,omitempty"`
}

// Zone defines model for Zone.
type Zone = dto.Zone

// ErrorResponse400 defines model for ErrorResponse400.
type ErrorResponse400 = Error

// ErrorResponse401 defines model for ErrorResponse401.
type ErrorResponse401 = Error

// ErrorResponse403 defines model for ErrorResponse403.
type ErrorResponse403 = Error

// ErrorResponse404 defines model for ErrorResponse404.
type ErrorResponse404 = Error

// ErrorResponse500 defines model for ErrorResponse500.
type ErrorResponse500 = Error

// ListJSONBody defines parameters for List.
type ListJSONBody struct {
	Paging   PagingRequest   `json:"paging,omitempty"`
	Searches []SearchRequest `json:"searches,omitempty"`
}

// ListJSONRequestBody defines body for List for application/json ContentType.
type ListJSONRequestBody ListJSONBody

// UpdateJSONRequestBody defines body for Update for application/json ContentType.
type UpdateJSONRequestBody = UpdateZoneClientRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List List all Zones of Client
	// (POST /api/v1/client/zones/search)
	List(c *gin.Context)
	// Get Get zone detail
	// (GET /api/v1/client/zones/{id})
	Get(c *gin.Context, id uint64)
	// Update Update zone
	// (PUT /api/v1/client/zones/{id})
	Update(c *gin.Context, id uint64)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// List operation middleware
func (siw *ServerInterfaceWrapper) List(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.List(c)
}

// Get operation middleware
func (siw *ServerInterfaceWrapper) Get(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id uint64

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "uint64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Get(c, id)
}

// Update operation middleware
func (siw *ServerInterfaceWrapper) Update(c *gin.Context) {

	var err error
	_ = err

	// ------------- Path parameter "id" -------------
	var id uint64

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "uint64", ValueIsUnescaped: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.Update(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/api/v1/client/zones/search", wrapper.List)
	router.GET(options.BaseURL+"/api/v1/client/zones/:id", wrapper.Get)
	router.PUT(options.BaseURL+"/api/v1/client/zones/:id", wrapper.Update)
}